+++ contracts/core/interfaces/IFeeCurve.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.15;
 
 interface IFeeCurve {
-    function currentRate(uint _percentageBorrowed) external view returns (uint);
+    function currentRate(uint256 _percentageBorrowed) external view returns (uint);
 }
